bugfix: code refactor
