+++ sdk/spring/spring-cloud-azure-autoconfigure/src/main/java/com/azure/spring/cloud/autoconfigure/redis/AzureJedisPasswordlessUtil.java
@@ -4,7 +4,7 @@
 package com.azure.spring.cloud.autoconfigure.redis;
 
 import com.azure.spring.cloud.autoconfigure.context.AzureGlobalProperties;
-import com.azure.spring.cloud.autoconfigure.implementation.passwordless.PasswordlessUtil;
+import com.azure.spring.cloud.core.implementation.util.AzurePropertiesUtils;
 import com.azure.spring.cloud.service.implementation.passwordless.AzureRedisPasswordlessProperties;
 import org.springframework.boot.autoconfigure.data.redis.RedisProperties;
 import org.springframework.boot.context.properties.PropertyMapper;
@@ -31,7 +31,10 @@ final class AzureJedisPasswordlessUtil {
     }
 
     static AzureRedisPasswordlessProperties mergeAzureProperties(AzureGlobalProperties azureGlobalProperties, AzureRedisPasswordlessProperties redisPasswordlessProperties) {
-        return (AzureRedisPasswordlessProperties) PasswordlessUtil.mergeAzureProperties(azureGlobalProperties, redisPasswordlessProperties);
+        AzureRedisPasswordlessProperties mergedProperties = new AzureRedisPasswordlessProperties();
+        AzurePropertiesUtils.mergeAzureCommonProperties(azureGlobalProperties, redisPasswordlessProperties, mergedProperties);
+        mergedProperties.setScopes(redisPasswordlessProperties.getScopes());
+        return mergedProperties;
     }
 
     static JedisClientConfiguration getJedisClientConfiguration(RedisProperties redisProperties) {
